Refactor transaction flow.
After writing out the docs, I realised that transactional operations involving multiple CRUD operations as provided by the storage are not possible. This is bad for the system because an extractor might panic while writing data to the database. If the service then restarts, the extractor will keep running into inconsistent state, continuing to panic.

To avoid this, transaction are now planned to be handled within use case specific gateways, which reuse the general purpose CRUD gateways. Only the use case can know exaclty the requirements of the transaction operation.