Compiling version of ExtractorRunner.
This helps highlighting where exactly there are problems with the Extractor trait. Had to make the trait generic to it is object-safe. Also had to move out the setup method else the trait wouldn't be object safe.

Bot Gateway and Message are not allowed to hold any references. For Message that is most likely fine, but for the Gateway we might want to be holding onto a references connection.