Improve get_account_delta.
We now return whether the account was created, deleted or just updated
as part of the delta messages.

There is an issue atm with the schema design that prevents us from
properly handle creation and deletion off accounts if it falls within
a forward/backward range. We would need to make the account table
versioned for this to work properly. There are workarounds possible with
the current schema as well though.

For now we only recognise the situation and error if we encounter it. As
most deltas will be for a single block it is unlikely that a contract
will be created and immediately deleted again or vice versa.