Improve decoding.
Instead of panicking we raise an error providing a nice error message in case the decoding of a literal type fails.

Note that we do not use the AbiDecode trait from ethers as this is meant to do decoding from solidity methods so it usualy required types to be 32byte padded. Additionally it is much more complex compated to our small custom trait.