ProtocolType: Don't add to PostgresGateway generic type.
Paraphrasing Louise's nicely-explained comment:

When you use the StorableProtocolType as a trait generic (what was being done before) you tell it that on initialisation it needs to set PT type as a struct that implements the StorableProtocolType traits. Instead, you can hardcode it to a suitable struct (ProtocolType) that we already know implements the StorableProtocolType trait.