Component Balances insertion: don't batch get tokens/txs/component_ids.
This needs to be more precise, otherwise tokens may be out of order, and we may not get one response for token that we query. It is much safer to loop through the component balances we want to insert, and search the db one by one for the corresponding item. Thus, nothing will be de-duplicated or out of order.