CachedGateway: Remove revert and flush functionality.
Initially tried to fix a bug here:
Sometimes it could happen that the cached gateway would complain about a missing link between pending and "to-be-inserted" block or block range. This is likely because we switched from single block insertions to emitting block ranges during syncs, so when we switch back into a sync, we may get a detached head.

Fixed another bug were component balances would be inserted with the wrong timestamps.

This commit finally removes revert and flush since we won't be using these anymore. They are too slow on a production db so we will change the revert mechanism to be in memory application side.