Simplify BlockSynchronizer.
It runs now. Tested with uniswap-v3 and v2 with tycho processing only finals blocks on mainnet. Messages are not aggregated anymore on the state synchronizer level. This makes synchronisation easier and removes a lot of overly complex code. Messages are then aggregated on the BlockSynchronizer when it makes sense (if there is a delay).

Some dead lock seems to happen occasionally if the channel buffers are very low (=1).