Increase DeltasClient buffer sizes.
Before these changes it was possible for the handle_message method to deadlock the websocket heartbeat messages. This could happen when the subscriber channel was full and the handle_message method would block until subscriber messages had been consumed. Any messages arriving to a full channel are now dropped, this is not ideal but prevents the application from deadlocks.