Remove protocol trait associated types.
This commit finishes a batch of simplifications to the storage trait.

Now it should be simple to factor out a separate crate for storage. This crate will use the newly created types.

In the future these types will also be used by the extractors since the blockchain specific types and validation should happen within the substreams.