Adjust migration after test on bigger dump.
Main change is that we now migrate latest entries ourselves, instead of using partition_data_proc. Reason being that the procedure would create a partition for our placeholder date in the year 262142. This is obviously not the desired behaviour, so we move these rows first, then afterward migrate the remaining data in the old table.

Reduced the migration period from 90 to 7d and instead we keep the old table around. This allows the migration to run quicker, reducing downtimes. If we need more historical data it can be accomplished by simply re-executing partition_data_proc.