Handle transaction lifecycle for update_token.
We create and commit a transaction for each update_tokens call directly since this is simpler and we don't need complex per block write cache handling for the token analysis cronjob.

Unfortunately this is not quite in line with the behaviour of other methods on the cached gateway therefore this is a short term fix and we should provide simpler gateway implementation without write caches.