Fixes panic when casting gas or tax to u64.
Some tokens set the balance of the receiver to the transfer amount or may expose a similar unexpected behaviour that causes tax to be unrealistically high. This case used to panic when casting tax to u64.

Although gas is usually capped to values that fit well into an u32 by RPCs. We apply a similar protection against overflow when casting to it to avoid panics.