fix(protocol_cache): Unkown shared tokens bug.
After actually sharing the protocol cache between extractors, it could happen that one extractor added a new token to the cache and another extractor then wrongly assumed the token has already been added to the database. This would crash the indexer with a NotFound error.

Additionally, if we omit tokens from emission, based on another extractor having witnessed it, then a client subscribed only to one (other) extractor might miss the token since the extractor will not emit it as new.

To fix both issues mentioned above, we'd need to store which extractor has seen a token. So the presence of a tokens is extractor specific. If an extractor has not seen the token yet, then it is added to the messages new tokens.

For now to keep complexity low, this fix simply emits all tokens as new tokens that appear as part of newly added components. This means some increased db communication overhead while inserting tokens but since token insertions should be rather rare it is the simplest solution for now.

Other approaches that were considered:

Simply retrying the transaction is not straightforward since we would have to check all tokens in a batch transaction for their actual db presence, then query the cache for the corresponding token instance.

Another workaround for this would be to share the cache with the DBCacheWriteExecutor, but that would mean a complex initialisation and adding complexity to the writer and the cache itself.