feat(otel): Improve otel config and span structure.
Otel resource attributes are now configurable via env variables.

Previously tycho would construct one big span per extractor. This would not play nicely with Grafana since this span quickly got way too big.

Instead, we construct a new root span now for each extractor loop. Unfortunately keeping the trace id as an attribute is now a bit more challenging so I decided to not include this in this commit. I'd evaluate first if this is necessary and in case it is we'd need to expose the SessionInit data via the BlockResponse struct.