fix(rpc): allow to run `RpcHandler` without pending deltas.
This is useful when running the `rpc` CLI command. In that case we don't run any extractor so it's expected that no `PendingDeltasBuffer` are registered in the RPC. But due to the recent changes in how we handle buffer lookups, it was failing if the buffer wasn't found.

This commit also refactor the service builder to avoid spawning unnecessary tasks and add docs.