fix(rpc): correctly handle when no ids are requested
This commit fixes a bug when no ids were given in a contract or protocol state request. In that case the current code would call the corresponding gateway function with `Some([])` for ids. Which would then lead to unexpected filtering on the query. We now pass down the `None` to the gateway function if no ids are requested.