fix: fix account balances bug
If an account is used for both a component and a token and the token is used outside of the protocol being deleted, it is skipped by the deletion script. This means we skip deleting account balances as well, which should be deleted as they're related to component accounts and not token accounts.